Update collider conditions, draw order, and tags in FP1LevelSpawn
The changes include updating the conditions for when a collider is needed based on object names in FP1LevelSpawn. The default draw order is set to -10000 from 0 (To prevent the player from being hidden by the environment). Also, specific tags are set for certain object names like 'Terrain Solid' and 'Terrain Platform'. The sorting order and layer assignment for foreground objects have also been altered. This brings the Fourth Screen of Dragon Valley pretty close to playable.